Unblocking barriers required by joining nodes
Joining nodes had two separate blocking points:
1. There is a blocking barrier in `_assign_worker_ranks()::_share_and_gather()`
	a. Uses `store_util.synchronize()` which requires comms with other workers
2. `_set_master_addr_port()` and `_get_master_addr_port()` require whichever nodes
  is rank 0 to set the master addr and port otherwise it will block

It is definitely an ugly hack. A lot of the functions already exist in
`project_pactum/agent/api.py` and I couldn't think of a clean way to make them
accessible to the EtcdRendezvousHandler in `etcd.py` so I just copied most of
them and I threw them in there.

Layer transfer still does not seem to be working after they start though,
so there are still bugs